Introduce AbstractRemoteTestNG
